Adding first version load balancer
Did a bit of Node.js research.
This is a pretty basic load balancer for me to work on and expand.Have
to say that I am amazed of it's power, with only 6 lines of code that
actually does stuff you get a basic routing system.
TODO:
- handle error (ex. one or more servers missing since I'm not sure how
it will react in that situation)
- try and implement some security , I need to know what sort of package
it will get from kannel to do that I guess.